Refactor BluetoothBackend + Devices screen
The devices screen and the bluetooth backend get modified in order
to make that the devices displayed depend on the backend's state,
making that the connected devices get displayed automatically
without the need of scanning the devices. This is not the case
for devices that are available but not connected, to display them
we must scan the devices, unless the device now disconnected was
previously displayed as connected (in case the user disconnected
it).