Integration dc page + backend
The Datacollection page now uses the backend with its states.

Unused static methods from the bluetooth backend are removed.

Also the Bluetooth backend is modified so that everytime a new
connection event happens, the mtu is required there, and also the
data collection and interpretation characteristics have their
notifications enabled, in order to decongest the communication
between app and devices when starting the data collections.